+++ packages/gatsby-remark-embed-snippet/src/index.js
@@ -4,8 +4,6 @@ const fs = require(`fs`)
 const normalizePath = require(`normalize-path`)
 const visit = require(`unist-util-visit`)
 
-const highlightCode = require(`gatsby-remark-prismjs/highlight-code`)
-
 // Language defaults to extension.toLowerCase();
 // This map tracks languages that don't match their extension.
 const FILE_EXTENSION_TO_LANGUAGE_MAP = {
@@ -13,6 +11,12 @@ const FILE_EXTENSION_TO_LANGUAGE_MAP = {
   md: `markup`,
   sh: `bash`,
   rb: `ruby`,
+  py: `python`,
+  ps1: `powershell`,
+  psm1: `powershell`,
+  bat: `batch`,
+  h: `c`,
+  tex: `latex`,
 }
 
 const getLanguage = file => {
@@ -27,10 +31,7 @@ const getLanguage = file => {
     : extension.toLowerCase()
 }
 
-module.exports = (
-  { markdownAST },
-  { classPrefix = `language-`, directory } = {}
-) => {
+module.exports = ({ markdownAST }, { directory } = {}) => {
   if (!directory) {
     throw Error(`Required option "directory" not specified`)
   } else if (!fs.existsSync(directory)) {
@@ -60,29 +61,10 @@ module.exports = (
       // @see https://github.com/PrismJS/prism/blob/1d5047df37aacc900f8270b1c6215028f6988eb1/themes/prism.css#L49-L54
       const language = getLanguage(file)
 
-      // Allow users to specify a custom class prefix to avoid breaking
-      // line highlights if Prism is required by any other code.
-      // This supports custom user styling without causing Prism to
-      // re-process our already-highlighted markup.
-      // @see https://github.com/gatsbyjs/gatsby/issues/1486
-      const className = language
-        .split(` `)
-        .map(token => `${classPrefix}${token}`)
-        .join(` `)
-
-      // Replace the node with the markup we need to make 100% width highlighted code lines work
-      try {
-        node.value = `<div class="gatsby-highlight">
-        <pre class="${className}"><code>${highlightCode(
-          language,
-          code
-        ).trim()}</code></pre>
-        </div>`
-        node.type = `html`
-      } catch (e) {
-        // rethrow error pointing to a file
-        throw Error(`${e.message}\nFile: ${file}`)
-      }
+      // Change the node type to code, insert our file as value and set language.
+      node.type = `code`
+      node.value = code
+      node.lang = language
     }
   })