+++ packages/gatsby-source-shopify/src/nodes.js
@@ -23,7 +23,7 @@ const { createNodeFactory, generateNodeId } = createNodeHelpers({
 
 const downloadImageAndCreateFileNode = async (
   { url, nodeId },
-  { createNode, createNodeId, touchNode, store, cache }
+  { createNode, createNodeId, touchNode, store, cache, reporter }
 ) => {
   let fileNodeID
 
@@ -43,6 +43,7 @@ const downloadImageAndCreateFileNode = async (
     createNode,
     createNodeId,
     parentNodeId: nodeId,
+    reporter,
   })
 
   if (fileNode) {
@@ -146,4 +147,4 @@ export const ProductVariantNode = imageArgs =>
 
 export const ShopPolicyNode = createNodeFactory(SHOP_POLICY)
 
-export const PagesNode = createNodeFactory(PAGE)
+export const PageNode = createNodeFactory(PAGE)