+++ packages/gatsby-telemetry/src/telemetry.js
@@ -1,13 +1,13 @@
-const { createHash } = require(`crypto`)
-const uuid = require(`uuid/v1`)
+const uuidv4 = require(`uuid/v4`)
 const EventStorage = require(`./event-storage`)
-const { sanitizeErrors } = require(`./error-helpers`)
+const { sanitizeErrors, cleanPaths } = require(`./error-helpers`)
 const ci = require(`ci-info`)
 const os = require(`os`)
-const { basename, join, sep } = require(`path`)
-const { execSync } = require(`child_process`)
+const { join, sep } = require(`path`)
 const isDocker = require(`is-docker`)
 const showAnalyticsNotification = require(`./showAnalyticsNotification`)
+const lodash = require(`lodash`)
+const { getRepositoryId } = require(`./repository-id`)
 
 module.exports = class AnalyticsTracker {
   store = new EventStorage()
@@ -17,15 +17,37 @@ module.exports = class AnalyticsTracker {
   osInfo // lazy
   trackingEnabled // lazy
   componentVersion
-  sessionId = uuid()
+  sessionId = this.getSessionId()
+
   constructor() {
     try {
       this.componentVersion = require(`../package.json`).version
       this.installedGatsbyVersion = this.getGatsbyVersion()
       this.gatsbyCliVersion = this.getGatsbyCliVersion()
+      this.defaultTags = this.getTagsFromEnv()
     } catch (e) {
       // ignore
     }
+  }
+
+  // We might have two instances of this lib loaded, one from globally installed gatsby-cli and one from local gatsby.
+  // Hence we need to use process level globals that are not scoped to this module
+  getSessionId() {
+    return (
+      process.gatsbyTelemetrySessionId ||
+      (process.gatsbyTelemetrySessionId = uuidv4())
+    )
+  }
+
+  getTagsFromEnv() {
+    if (process.env.GATSBY_TELEMETRY_TAGS) {
+      try {
+        return JSON.parse(process.env.GATSBY_TELEMETRY_TAGS)
+      } catch (_) {
+        // ignore
+      }
+    }
+    return {}
   }
 
   getGatsbyVersion() {
@@ -60,7 +82,7 @@ module.exports = class AnalyticsTracker {
     }
     return undefined
   }
-  captureEvent(type = ``, tags = {}) {
+  captureEvent(type = ``, tags = {}, opts = { debounce: false }) {
     if (!this.isTrackingEnabled()) {
       return
     }
@@ -71,25 +93,56 @@ module.exports = class AnalyticsTracker {
     }
 
     const decoration = this.metadataCache[type]
-    delete this.metadataCache[type]
     const eventType = `${baseEventType}_${type}`
-    this.buildAndStoreEvent(eventType, Object.assign(tags, decoration))
+
+    if (opts.debounce) {
+      const debounceTime = 5 * 1000
+      const now = Date.now()
+      const debounceKey = JSON.stringify({ type, decoration, tags })
+      const last = this.debouncer[debounceKey] || 0
+      if (now - last < debounceTime) {
+        return
+      }
+      this.debouncer[debounceKey] = now
+    }
+
+    delete this.metadataCache[type]
+    this.buildAndStoreEvent(eventType, lodash.merge({}, tags, decoration))
   }
 
   captureError(type, tags = {}) {
     if (!this.isTrackingEnabled()) {
       return
     }
+
     const decoration = this.metadataCache[type]
     delete this.metadataCache[type]
     const eventType = `CLI_ERROR_${type}`
 
     if (tags.error) {
       // `error` ought to have been `errors` but is `error` in the database
-      tags.error = sanitizeErrors(tags.error)
+      if (Array.isArray(tags.error)) {
+        const { error, ...restOfTags } = tags
+        error.forEach(err => {
+          this.captureError(type, { error: err, ...restOfTags })
+        })
+        return
+      }
+
+      tags.errorV2 = {
+        id: tags.error.id,
+        text: cleanPaths(tags.error.text),
+        level: tags.error.level,
+        type: tags.error?.type,
+        // see if we need empty string or can just use NULL
+        stack: cleanPaths(tags.error?.error?.stack || ``),
+        context: cleanPaths(JSON.stringify(tags.error?.context)),
+      }
+
+      delete tags.error
     }
 
-    this.buildAndStoreEvent(eventType, Object.assign(tags, decoration))
+    this.buildAndStoreEvent(eventType, lodash.merge({}, tags, decoration))
   }
 
   captureBuildError(type, tags = {}) {
@@ -105,23 +158,22 @@ module.exports = class AnalyticsTracker {
       tags.error = sanitizeErrors(tags.error)
     }
 
-    this.buildAndStoreEvent(eventType, Object.assign(tags, decoration))
+    this.buildAndStoreEvent(eventType, lodash.merge({}, tags, decoration))
   }
 
   buildAndStoreEvent(eventType, tags) {
     const event = {
       installedGatsbyVersion: this.installedGatsbyVersion,
       gatsbyCliVersion: this.gatsbyCliVersion,
-      ...this.defaultTags,
-      ...tags, // The schema must include these
+      ...lodash.merge({}, this.defaultTags, tags), // The schema must include these
       eventType,
       sessionId: this.sessionId,
       time: new Date(),
       machineId: this.getMachineId(),
-      repositoryId: this.getRepoId(),
       componentId: `gatsby-cli`,
       osInformation: this.getOsInfo(),
       componentVersion: this.componentVersion,
+      ...getRepositoryId(),
     }
     this.store.addEvent(event)
   }
@@ -133,7 +185,7 @@ module.exports = class AnalyticsTracker {
     }
     let machineId = this.store.getConfig(`telemetry.machineId`)
     if (!machineId) {
-      machineId = uuid()
+      machineId = uuidv4()
       this.store.updateConfig(`telemetry.machineId`, machineId)
     }
     this.machineId = machineId
@@ -155,25 +207,6 @@ module.exports = class AnalyticsTracker {
     }
     this.trackingEnabled = enabled
     return enabled
-  }
-
-  getRepoId() {
-    // we may live multiple levels in git repo
-    let prefix = `pwd:`
-    let repo = basename(process.cwd())
-    try {
-      const originBuffer = execSync(
-        `git config --local --get remote.origin.url`,
-        { timeout: 1000, stdio: `pipe` }
-      )
-      repo = String(originBuffer).trim()
-      prefix = `git:`
-    } catch (e) {
-      // ignore
-    }
-    const hash = createHash(`sha256`)
-    hash.update(repo)
-    return prefix + hash.digest(`hex`)
   }
 
   getOsInfo() {