+++ www/src/components/shared/footer-links.js
@@ -8,22 +8,23 @@ const FooterList = styled.ul`
   border-top: 1px solid ${colors.ui.border.subtle};
   font-size: ${fontSizes[1]};
   margin: 0;
-  padding: 0;
-  padding-top: ${space[9]};
+  padding: ${space[9]} ${space[6]};
   margin-top: ${space[9]};
-  padding-bottom: ${space[9]};
   list-style: none;
   text-align: center;
   width: 100%;
 
   li {
     display: inline-block;
-
-    &:first-of-type:after {
+    &:after {
       color: ${colors.grey[30]};
       content: "•";
       padding-left: 1em;
-      margin-right: 1em;
+      padding-right: 1em;
+    }
+
+    &:last-of-type:after {
+      content: none;
     }
 
     a {
@@ -42,6 +43,9 @@ const FooterLinks = props => (
   <FooterList bottomMargin={props.bottomMargin}>
     <li>
       <Link to="/accessibility-statement">Accessibility Statement</Link>
+    </li>
+    <li>
+      <a href="/guidelines/logo/">Logo &amp; Assets</a>
     </li>
     <li>
       <a href="https://www.gatsbyjs.com">Gatsbyjs.com</a>