+++ packages/babel-plugin-remove-graphql-queries/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.1](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.7.0...babel-plugin-remove-graphql-queries@2.7.1) (2019-07-11)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
+# [2.7.0](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.6.3...babel-plugin-remove-graphql-queries@2.7.0) (2019-06-20)
+
+**Note:** Version bump only for package babel-plugin-remove-graphql-queries
+
 ## [2.6.3](https://github.com/gatsbyjs/gatsby/compare/babel-plugin-remove-graphql-queries@2.6.2...babel-plugin-remove-graphql-queries@2.6.3) (2019-04-02)
 
 ### Features