+++ packages/gatsby-cli/CHANGELOG.md
@@ -3,6 +3,116 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-cli@2.7.9...gatsby-cli@2.7.10) (2019-07-11)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.7.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.8...gatsby-cli@2.7.9) (2019-07-10)
+
+### Bug Fixes
+
+- don't break joi validation for production bundles webpack… ([#15602](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15602)) ([4c50024](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/4c50024))
+- don't break joi validation for unhandledRejections & apirunner ([#15600](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15600)) ([14ba538](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/14ba538))
+
+## [2.7.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.7...gatsby-cli@2.7.8) (2019-07-05)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.7.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.6...gatsby-cli@2.7.7) (2019-06-29)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.7.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.5...gatsby-cli@2.7.6) (2019-06-28)
+
+### Bug Fixes
+
+- **gatsby-cli:** depend upon [@hapi](https://github.com/hapi)/joi ([#15221](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15221)) ([59a021d](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/59a021d))
+
+## [2.7.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.4...gatsby-cli@2.7.5) (2019-06-28)
+
+### Features
+
+- **gatsby-cli:** Add error codes and structured errors ([#14904](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14904)) ([d26651e](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/d26651e))
+
+## [2.7.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.3...gatsby-cli@2.7.4) (2019-06-28)
+
+### Bug Fixes
+
+- filter out frames that are not matched to source location ([#15211](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15211)) ([44e8f76](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/44e8f76))
+
+## [2.7.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.2...gatsby-cli@2.7.3) (2019-06-27)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.7.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.1...gatsby-cli@2.7.2) (2019-06-24)
+
+### Bug Fixes
+
+- **gatsby-cli:** Added better-opn as a dependency ([#15078](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15078)) ([d855e50](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/d855e50))
+
+### Features
+
+- **gatsby-cli:** improve version output ([#14924](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14924)) ([6210ec3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/6210ec3))
+
+## [2.7.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.7.0...gatsby-cli@2.7.1) (2019-06-24)
+
+### Bug Fixes
+
+- **gatsby:** minor refactor of console functions ([#14956](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14956)) ([7775b3e](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/7775b3e))
+- cleanup stack traces for html build errors ([#15050](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/15050)) ([d029f7b](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/d029f7b))
+
+### Features
+
+- **gatsby-cli:** Prompt user for options (no args with gatsby new) ([#14097](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14097)) ([99fb7b4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/99fb7b4))
+
+# [2.7.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.13...gatsby-cli@2.7.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.6.13](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.12...gatsby-cli@2.6.13) (2019-06-19)
+
+### Bug Fixes
+
+- **gatsby-cli:** add missing node-fetch dependency ([#14908](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14908)) ([b7da1e4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/b7da1e4))
+
+## [2.6.12](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.9...gatsby-cli@2.6.12) (2019-06-19)
+
+### Bug Fixes
+
+- fix gatsby-cli dep in source-filesystem & plugin-sharp ([#14881](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14881)) ([2594623](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/2594623))
+
+## [2.6.11](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.10...gatsby-cli@2.6.11) (2019-06-19)
+
+### Bug Fixes
+
+- fix gatsby-cli dep in source-filesystem & plugin-sharp ([#14881](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14881)) ([2594623](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/2594623))
+
+## [2.6.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.9...gatsby-cli@2.6.10) (2019-06-18)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.6.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.8...gatsby-cli@2.6.9) (2019-06-18)
+
+**Note:** Version bump only for package gatsby-cli
+
+## [2.6.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.7...gatsby-cli@2.6.8) (2019-06-18)
+
+### Features
+
+- **gatsby-cli:** move progressbar into ink ([#14220](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14220)) ([967597c](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/967597c))
+
+## [2.6.7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.5...gatsby-cli@2.6.7) (2019-06-12)
+
+### Bug Fixes
+
+- **gatsby-cli:** create spawnWithArgs to be able to call spawn with command containing spaces ([#14698](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14698)) ([7b3efe7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/7b3efe7))
+
+## [2.6.6](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.5...gatsby-cli@2.6.6) (2019-06-12)
+
+### Bug Fixes
+
+- **gatsby-cli:** create spawnWithArgs to be able to call spawn with command containing spaces ([#14698](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/issues/14698)) ([7b3efe7](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/commit/7b3efe7))
+
 ## [2.6.5](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-cli/compare/gatsby-cli@2.6.4...gatsby-cli@2.6.5) (2019-06-05)
 
 ### Features