+++ packages/gatsby-plugin-layout/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.1.0...gatsby-plugin-layout@1.1.1) (2019-07-11)
+
+**Note:** Version bump only for package gatsby-plugin-layout
+
+# [1.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-layout/compare/gatsby-plugin-layout@1.0.15...gatsby-plugin-layout@1.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-plugin-layout
+
 ## [1.0.15](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-layout/compare/gatsby-plugin-layout@1.0.14...gatsby-plugin-layout@1.0.15) (2019-04-23)
 
 **Note:** Version bump only for package gatsby-plugin-layout