+++ packages/gatsby-source-mongodb/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-mongodb@2.1.1...gatsby-source-mongodb@2.1.2) (2019-07-11)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
+## [2.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.1.0...gatsby-source-mongodb@2.1.1) (2019-07-02)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
+# [2.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.0.20...gatsby-source-mongodb@2.1.0) (2019-06-20)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
 ## [2.0.20](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.0.19...gatsby-source-mongodb@2.0.20) (2019-06-10)
 
 ### Bug Fixes